requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
bcrypt==4.2.0
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2